refactor: move common functions to utils module
