+++ web/frontend/tests/integration/utils.ts
@@ -1,3 +1,4 @@
+import { test, expect } from '@playwright/test';
 import { default as i18n } from 'i18next';
 import en from './../../src/language/en.json';
 import de from './../../src/language/de.json';
@@ -27,4 +28,21 @@ export async function logOut (page: any) {
   await page.getByRole('button', { name: i18n.t('Profile') }).dispatchEvent('click');
   await page.getByRole('menuitem', { name: i18n.t('logout') }).dispatchEvent('click');
   await page.getByRole('button', { name: i18n.t('continue') }).dispatchEvent('click');
+}
+
+export async function assertOnlyVisibleToAuthenticated(page: any, role: string, key: string) {
+  const element = page.getByRole(role, { name: i18n.t(key) });
+  await expect(element).toBeHidden();   // assert is hidden to unauthenticated user
+  await logInNonAdmin(page);
+  await expect(element).toBeVisible();  // assert is visible to authenticated non-admin user
+}
+
+export async function assertOnlyVisibleToAdmin(page: any, role: string, key: string) {
+  const element = page.getByRole(role, { name: i18n.t(key) });
+  await expect(element).toBeHidden();   // assert is hidden to unauthenticated user
+  await logInNonAdmin(page);
+  await expect(element).toBeHidden();   // assert is hidden to non-admin user
+  await logOut(page);
+  await logIn(page);
+  await expect(element).toBeVisible();  // assert is visible to admin user
 }
